requirements.txt
SQLAlchemy
alembic
pycapnp
pycrypto
pyyaml
-e git://github.com/tonysimpson/nanomsg-python.git#egg=nanomsg
